+++ templates/settings-admin.php
@@ -1,0 +1,1 @@
+<div id="gitcloud-settings-admin"></div>