+++ templates/settings-personal.php
@@ -1,0 +1,1 @@
+<div id="gitcloud-settings-personal"></div>